fix: ability to reference tables
this involved a literal saga, where I had to package a patch of a python-based pandoc filter called pandoc-xnos and the subsequent packages it supports, in order to use the patch. This patch allowed for pandoc version 3 by fixing a single character in a stupid regex. This is not how software should be. Nevertheless, it works, and it's reproducable, which is what matters most. In order to use the pandoc-secnos package, however, I will need to fork my own copy of it that avoids checking to see if its own pakage name exists in the PATH. This is the only one of the guy's packages that does it, and I have no idea why.